refactor: another approach for organizing the custom functions
