+++ libs/safe/src/lib/survey/functions/index.ts
@@ -1,0 +1,26 @@
+import addDays from './addDays';
+import getAidsGiven from './getAidsGiven';
+import getMatrixTitles from './getMatrixTitles';
+import getTotalAids from './getTotalAids';
+import intersect from './intersect';
+import length from './length';
+import listColsForRows from './listColsForRows';
+import listRowsWithColValue from './listRowsWithColValue';
+import nextYear from './nextYear';
+import nl2br from './nl2br';
+import weekday from './weekday';
+
+/** Custom functions used  */
+export const functions = [
+  getTotalAids,
+  nextYear,
+  weekday,
+  getAidsGiven,
+  addDays,
+  listRowsWithColValue,
+  listColsForRows,
+  nl2br,
+  getMatrixTitles,
+  length,
+  intersect,
+];
