+++ libs/safe/src/lib/survey/functions/nextYear.ts
@@ -1,0 +1,6 @@
+/** @returns a date object set to next year */
+export default () => {
+  const date = new Date();
+  date.setFullYear(date.getFullYear() + 1);
+  return date;
+};
