+++ libs/safe/src/lib/survey/init.ts
@@ -7,7 +7,7 @@ import { UntypedFormBuilder } from '@angular/forms';
 import { DomService } from '../services/dom/dom.service';
 import { SafeAuthService } from '../services/auth/auth.service';
 import { SafeReferenceDataService } from '../services/reference-data/reference-data.service';
-import addCustomFunctions from '../utils/custom-functions';
+import addCustomFunctions from './custom-functions';
 import * as ResourceComponent from './components/resource';
 import * as ResourcesComponent from './components/resources';
 import * as OwnerComponent from './components/owner';
